requirements.txt
# Core dependencies for SSIS to PySpark conversion
openai>=1.0.0
google-generativeai>=0.3.0
lxml>=4.9.0
xmltodict>=0.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
click>=8.0.0
rich>=13.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# For building .exe
pyinstaller>=5.0.0
